integrate all api

+++ backend/src/cart/cart.service.ts
@@ -21,18 +21,73 @@ export class CartService {
       });
   }
 
-  async deleteCart(cartId: number) {
+  async deleteCart(userId: number, cartId: number) {
+    const cartItem = await this.prisma.cart.findFirst({
+      where: { id: cartId, userId: userId },
+    });
+
+    if (!cartItem) {
+      throw new Error('Not found Item');
+    }
+
     return this.prisma.cart.delete({
       where: { id: cartId },
     });
   }
 
   async getUserCart(userId: number) {
-    return this.prisma.cart.findMany({
+    const cartData = await this.prisma.cart.findMany({
       where: { userId },
       include: {
-        product: true, // ดึงข้อมูลสินค้าใน Cart
+        product: true,
       },
+      orderBy: {
+        createdAt: 'desc',
+      },
+    });
+
+    const totalQuantity = cartData.reduce(
+      (sum, item) => sum + item.quantity,
+      0,
+    );
+
+    const totalPrice = +cartData
+      .reduce((sum, item) => sum + item.product.price * item.quantity, 0)
+      .toFixed(2);
+
+    return {
+      cartData: cartData,
+      length: (await cartData).length,
+      totalQuantity: totalQuantity,
+      totalPrice: totalPrice,
+    };
+  }
+
+  async updateCart(
+    userId: number,
+    cartId: number,
+    action: 'increment' | 'decrement',
+  ) {
+    const cartItem = await this.prisma.cart.findFirst({
+      where: { id: cartId, userId: userId },
+    });
+
+    if (!cartItem) {
+      throw new Error('Cart item not found or does not belong to this user');
+    }
+
+    const newQuantity =
+      action === 'increment' ? cartItem.quantity + 1 : cartItem.quantity - 1;
+
+    if (newQuantity < 1) {
+      return this.prisma.cart.delete({
+        where: { id: cartId },
+      });
+    }
+
+    return this.prisma.cart.update({
+      where: { id: cartId },
+      data: { quantity: newQuantity },
     });
   }
 }
